// models/report_aas.ts
export interface ReportAAS {
  id: number;
  type: string;
  type_text: string;
  emotional_stability_level: number;
  overall_evaluation: string;
  detail_evaluations: string;
  counseling_text: string;
  emotional_stability_text: string;
  title: string;
  sequence: number;
  created_at: string;
  updated_at: string;
}